feat: Remove regra de negócio da views.py e centraliza em service/service.py
